increase decrease 해결 및 email 아직 불완전

+++ backend/src/main/java/triplej/capstone/controllers/CurrentTableController.java
@@ -14,12 +14,12 @@ public class CurrentTableController {
     CurrentTableService currentTableService;
 
     @GetMapping("/increment/{category_id}/{restaurant_id}")
-    public CurrentTableDto increaseCurrentTable(@PathVariable("category_id") int category_id, @PathVariable("restaurant_id") int id){
+    public CurrentTableDto increaseCurrentTable(@PathVariable("category_id") int category_id, @PathVariable("restaurant_id") long id){
         return  currentTableService.increment(category_id, id);
     }
 
     @GetMapping("/decrement/{category_id}/{restaurant_id}")
-    public CurrentTableDto decreaseCurrentTable(@PathVariable("category_id") int category_id, @PathVariable("restaurant_id") int id){
+    public CurrentTableDto decreaseCurrentTable(@PathVariable("category_id") int category_id, @PathVariable("restaurant_id") long id){
         return  currentTableService.decrement(category_id, id);
     }
